fix(#1501): move Remote exception types from remote/client to core/exceptions
Eliminates cross-layer coupling where fuse/operations.py imported
RemoteFilesystemError, RemoteConnectionError, and RemoteTimeoutError
from nexus.remote.client. These exception types are now defined in
core/exceptions (the canonical location for all Nexus exceptions),
allowing both fuse and remote layers to import without cross-layer
dependency. Also removes the defensive try/except ImportError pattern
in fuse/operations.py since core.exceptions is always available.
