+++ src/nexus/remote/__init__.py
@@ -31,14 +31,16 @@ Example (async):
     ...     await nx.skills.create("my-skill", "A skill", template="basic")
 """
 
+from nexus.core.exceptions import (
+    RemoteConnectionError,
+    RemoteFilesystemError,
+    RemoteTimeoutError,
+)
 from nexus.remote.async_client import (
     AsyncRemoteNexusFS,
 )
 from nexus.remote.client import (
-    RemoteConnectionError,
-    RemoteFilesystemError,
     RemoteNexusFS,
-    RemoteTimeoutError,
 )
 from nexus.remote.domain import (
     AsyncACEClient,
